lib-storage: raw-mail - Cast st_size to uoff_t
mbox->size is unsigned field, and st_size will be -1
is the stat target is not regular file or symbolic link.
This can happen if it's stream like stdin.

+++ src/lib-storage/index/raw/raw-mail.c
@@ -31,7 +31,7 @@ static int raw_mail_stat(struct mail *mail)
 		mbox->mtime = st->st_mtime;
 	if (mbox->ctime != (time_t)-1)
 		mbox->ctime = st->st_ctime;
-	mbox->size = st->st_size;
+	mbox->size = (size_t)st->st_size;
 	return 0;
 }
